Выносит функцию для рендера в отдельный файл
