+++ src/utils.js
@@ -1,9 +1,3 @@
-// Словарь позиций для вставки
-const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`,
-};
-
 // Функция форматирования времени
 const castTimeFormat = (value) => {
   return value < 10 ? `0${value}` : String(value);
@@ -16,24 +10,4 @@ const formatTime = (date) => {
   return `${hours}:${minutes}`;
 };
 
-// Функция создания элемента
-const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-// Функция рендера элементов
-const render = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
-
-export {RenderPosition, formatTime, createElement, render};
+export {formatTime};
